feat Product 테이블에 createdBy 컬럼 추가, 상품 수정 삭제 작성자만 할 수 있게

+++ src/main/java/com/basic_shop/shop/service/ProductService.java
@@ -4,6 +4,8 @@ import com.basic_shop.shop.dto.ProductDto;
 import com.basic_shop.shop.entity.Product;
 import com.basic_shop.shop.repository.ProductRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -21,10 +23,11 @@ public class ProductService {
     }
 
     //상품 등록
-    public void addProduct(ProductDto productDto) {
+    public void addProduct(ProductDto productDto, Authentication authentication) {
         Product product = new Product();
         product.setTitle(productDto.getTitle());
         product.setPrice(productDto.getPrice());
+        product.setCreatedBy(authentication.getName());
         productRepository.save(product);
     }
 
@@ -45,21 +48,30 @@ public class ProductService {
         productDto.setId(product.getId());
         productDto.setTitle(product.getTitle());
         productDto.setPrice(product.getPrice());
+        productDto.setCreatedBy(product.getCreatedBy());
         return productDto;
     }
 
     //상품 수정
-    public void updatePost(Long id, ProductDto productDto) {
+    public void updatePost(Long id, ProductDto productDto, Authentication authentication) {
         Product product = getProduct(id);
-        product.setId(id);
-        product.setTitle(productDto.getTitle());
-        product.setPrice(productDto.getPrice());
-        productRepository.save(product);
+        if (authentication.getName().equals(product.getCreatedBy())) {
+            product.setId(id);
+            product.setTitle(productDto.getTitle());
+            product.setPrice(productDto.getPrice());
+            productRepository.save(product);
+        }else {
+            throw new AccessDeniedException("Only the author can edit this post.");
+        }
     }
 
     //상품 삭제
-    public void deletePost(Long id) {
+    public void deletePost(Long id, Authentication authentication) {
         Product product = getProduct(id);
-        productRepository.delete(product);
+        if (authentication.getName().equals(product.getCreatedBy())) {
+            productRepository.delete(product);
+        }else{
+            throw new AccessDeniedException("Only the author can delete this post.");
+        }
     }
 }
